+++ packages/graphql/tests/integration/subscriptions/create/create-authorization.int.test.ts
@@ -69,9 +69,6 @@ describe("auth/bind", () => {
                             create: [{
                                 node: {
                                     id: "post-id-1",
-                                    creator: {
-                                        create: { node: {id: "not valid"} }
-                                    }
                                 }
                             }]
                         }