+++ packages/graphql/tests/tck/directives/authorization/arguments/bind/bind.test.ts
@@ -400,7 +400,7 @@ describe("Cypher Auth Allow", () => {
         expect(formatCypher(result.cypher)).toMatchInlineSnapshot(`
             "MATCH (this:Post)
             WHERE this.id = $param0
-            WITH this
+            WITH *
             CALL {
             	WITH this
             	OPTIONAL MATCH (this_connect_creator0_node:User)
@@ -421,7 +421,7 @@ describe("Cypher Auth Allow", () => {
             WITH *, count(authorization_this0) AS creatorCount
             WITH *
             WHERE (apoc.util.validatePredicate(NOT ($isAuthenticated = true AND (creatorCount <> 0 AND ($jwt.sub IS NOT NULL AND authorization_this0.id = $jwt.sub))), \\"@neo4j/graphql/FORBIDDEN\\", [0]) AND apoc.util.validatePredicate(NOT ($isAuthenticated = true AND ($jwt.sub IS NOT NULL AND this_connect_creator0_node.id = $jwt.sub)), \\"@neo4j/graphql/FORBIDDEN\\", [0]))
-            	RETURN count(*) AS connect_this_connect_creator_User
+            	RETURN count(*) AS connect_this_connect_creator_User0
             }
             WITH *
             WITH *