+++ packages/graphql/tests/tck/issues/583.test.ts
@@ -87,12 +87,12 @@ describe("https://github.com/neo4j/graphql/issues/583", () => {
                 CALL {
                     WITH *
                     MATCH (this)-[this0:ACTED_IN]->(this1:Movie)
-                    WITH this1 { .awardsGiven, .title, __resolveType: \\"Movie\\", __id: id(this) } AS this1
+                    WITH this1 { .title, .awardsGiven, __resolveType: \\"Movie\\", __id: id(this) } AS this1
                     RETURN this1 AS var2
                     UNION
                     WITH *
                     MATCH (this)-[this3:ACTED_IN]->(this4:Series)
-                    WITH this4 { .awardsGiven, .title, __resolveType: \\"Series\\", __id: id(this) } AS this4
+                    WITH this4 { .title, .awardsGiven, __resolveType: \\"Series\\", __id: id(this) } AS this4
                     RETURN this4 AS var2
                     UNION
                     WITH *